+++ client/stores/appStore.ts
@@ -1,3 +1,5 @@
+/*stores/appStore.ts
+@Detail:用户的初始数据，未登录时，需要有一个默认名*/
 import { create } from 'zustand';
 
 interface AppState {