+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 3. Post page and Dashboard welcome words
 4. delete AI chat(use Deepseek Agent)
 5. unused: chart page
+6. better sidebar style
+7. better module config
+8. personalization page
 
 ### 0.4.2
 1. delete "Process manager" module